backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
pydantic==2.7.1
